Safari-611.1.10/LayoutTests/editing/undo-manager/undo-manager-add-item-exceptions-expected.txt

Verifies that UndoManager.addItem() throws JavaScript exceptions. This test requires WebKitTestRunner.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS firstFrameDocument.undoManager.addItem(item) threw exception SecurityError: A browsing context is required to add an UndoItem.
PASS document.undoManager.addItem(item) did not throw exception.
PASS secondFrameDocument.undoManager.addItem(item) threw exception InvalidModificationError: This item has already been added to an UndoManager.
Performed undo.
Performed redo.
PASS successfullyParsed is true

TEST COMPLETE
